telegram_service/requirements.txt
# Telegram Client
Telethon>=1.24.0

# Web Framework
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0

# Database
sqlalchemy>=1.4.0
aiomysql>=0.1.14

# Utilities
python-dotenv>=0.19.0
aiofiles>=0.7.0
Pillow>=8.3.0

# For async operations
asyncio>=3.4.3
aiohttp>=3.7.4

# For testing
pytest>=6.2.4